Allow refresh of views data

+++ odk_parser.py
@@ -537,7 +537,7 @@ class OdkParser():
 
         # now we have all our selected submissions as csv files, so process them
         try:
-            table_views = self.save_csvs_to_database(full_path, prop_view_name)
+            table_views = self.save_csvs_to_database(full_path, prop_view_name, repopulate)
         except Exception:
             raise
 
@@ -565,6 +565,8 @@ class OdkParser():
 
             # save these submissions to the database
             terminal.tprint("\tSaving the views extracted submissions", 'okblue')
+            if repopulate:
+                ViewsData.objects.filter(view=form_view).delete()
             for submission in all_submissions:
                 new_submission = ViewsData(
                     view=form_view,
@@ -629,9 +631,11 @@ class OdkParser():
             # raise Exception("Duplicate view name '%s'. Can't save." % view_name)
             # return
 
-    def save_csvs_to_database(self, full_path, prop_view_name):
+    def save_csvs_to_database(self, full_path, prop_view_name, repopulate):
+        is_create_table = '--no-create' if repopulate else ''
+
         # import_command = "csvsql --db 'postgresql:///%s?user=%s&password=%s' --encoding utf-8 --blanks --insert --tables %s %s"
-        import_command = "env/bin/csvsql --db '%s:///%s?user=%s&password=%s' --encoding utf8 --blanks --no-inference --insert --tables %s %s"
+        import_command = "env/bin/csvsql --db '%s:///%s?user=%s&password=%s' --encoding utf8 --blanks %s --no-inference --insert --tables %s %s"
         terminal.tprint(import_command, 'warn')
         table_views = []
         for filename in os.listdir(full_path):
@@ -657,11 +661,17 @@ class OdkParser():
                     settings.DATABASES['default']['NAME'],
                     settings.DATABASES['default']['USER'],
                     settings.DATABASES['default']['PASSWORD'],
+                    is_create_table,
                     table_name_hash_dig,
                     filename.decode('utf-8'),
                 )
 
                 try:
+                    if repopulate:
+                        # truncate the table inorder to add new data
+                        with connection.cursor() as cursor:
+                            dquery = "truncate table %s" % table_name_hash_dig
+                            cursor.execute(dquery)
                     terminal.tprint("\tRunning the command '%s'" % cmd, 'ok')
                     # run commands to create primary key
                     print(subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE).stdout.read())
@@ -671,37 +681,44 @@ class OdkParser():
                     raise Exception("Error while converting the csv file '%s' to a table '%s'." % (filename, table_name))
 
                 try:
-                    with connection.cursor() as cursor:
-                        logging.debug("Adding a primary key constraint for the table '%s'" % table_name)
-                        query = "alter table %s add primary key (%s)" % (table_name_hash_dig, 'unique_id')
-                        cursor.execute(query)
-
-                        # if table name has a main on it, it must have a _uuid field which should be unique
-                        if re.search("main$", table_name) is not None:
-                            # this is finicky, omit it for now
-                            terminal.tprint("Not adding a unique constraint for column '_uuid'", 'fail')
-                            # logging.debug("Adding unique constraint '%s' for the table '%s'" % ('_uuid', table_name))
-                            # uquery = "alter table %s add constraint %s_%s unique (%s)" % (table_name_hash_dig, table_name_hash_dig, 'uuid', '_uuid')
-                            # cursor.execute(uquery)
-                        else:
-                            # for the other tables, add an index to top_id
-                            logging.debug("Adding indexes to '%s' and '%s' for the table '%s'" % ('top_id', 'parent_id', table_name))
-                            uquery = "create index %s_%s on %s (%s)" % (table_name_hash_dig, 'top_id', table_name_hash_dig, 'top_id')
-                            cursor.execute(uquery)
-                            uquery = "create index %s_%s on %s (%s)" % (table_name_hash_dig, 'parent_id', table_name_hash_dig, 'parent_id')
-                            cursor.execute(uquery)
-                except Exception as e:
-                    logging.error("For some reason can't create a primary key or unique key, raise an error and delete the view")
-                    terminal.tprint(str(e), 'fail')
-                    with connection.cursor() as cursor:
-                        dquery = "drop table %s" % table_name_hash_dig
-                        cursor.execute(dquery)
-                    sentry.captureException()
-                    raise Exception("For some reason I can't create a primary key or unique key for the table %s. Deleting it entirely" % table_name)
-
+                    if repopulate is False:
+                        self.add_dynamic_table_keys(table_name, table_name_hash_dig, repopulate)
+                except Exception:
+                    raise
                 table_views.append({'table_name': table_name, 'hashed_name': table_name_hash_dig})
 
         return table_views
+
+    def add_dynamic_table_keys(self, table_name, table_name_hash_dig, repopulate):
+        try:
+            with connection.cursor() as cursor:
+                logging.debug("Adding a primary key constraint for the table '%s'" % table_name)
+                query = "alter table %s add primary key (%s)" % (table_name_hash_dig, 'unique_id')
+                cursor.execute(query)
+
+                # if table name has a main on it, it must have a _uuid field which should be unique
+                if re.search("main$", table_name) is not None:
+                    # this is finicky, omit it for now
+                    terminal.tprint("Not adding a unique constraint for column '_uuid'", 'fail')
+                    # logging.debug("Adding unique constraint '%s' for the table '%s'" % ('_uuid', table_name))
+                    # uquery = "alter table %s add constraint %s_%s unique (%s)" % (table_name_hash_dig, table_name_hash_dig, 'uuid', '_uuid')
+                    # cursor.execute(uquery)
+                else:
+                    # for the other tables, add an index to top_id
+                    logging.debug("Adding indexes to '%s' and '%s' for the table '%s'" % ('top_id', 'parent_id', table_name))
+                    uquery = "create index %s_%s on %s (%s)" % (table_name_hash_dig, 'top_id', table_name_hash_dig, 'top_id')
+                    cursor.execute(uquery)
+                    uquery = "create index %s_%s on %s (%s)" % (table_name_hash_dig, 'parent_id', table_name_hash_dig, 'parent_id')
+                    cursor.execute(uquery)
+        except Exception as e:
+            logging.error("For some reason can't create a primary key or unique key, raise an error and delete the view")
+            terminal.tprint(str(e), 'fail')
+            if repopulate is False:
+                with connection.cursor() as cursor:
+                    dquery = "drop table %s" % table_name_hash_dig
+                    cursor.execute(dquery)
+            sentry.captureException()
+            raise Exception("For some reason I can't create a primary key or unique key for the table %s. Deleting it entirely" % table_name)
 
     def formulate_view_name(self, view_name, form_group):
         """
@@ -800,7 +817,7 @@ class OdkParser():
                 return all_submissions
 
         # check if there is need to create a database view of this data
-        if download_type == 'download_save':
+        if download_type == 'download_save' or update_local_data:
             try:
                 self.save_user_view(form_id, view_name, nodes, all_submissions, self.output_structure, form_group.group_name, update_local_data)
             except Exception as e:
@@ -816,6 +833,8 @@ class OdkParser():
             output_name = './' + form_name + '_' + now + '.xlsx'
             self.save_submissions_as_excel(all_submissions, self.output_structure, output_name)
             return {'is_downloadable': True, 'filename': output_name}
+        else:
+            return all_submissions
 
     def save_submissions_as_excel(self, submissions, structure, filename):
         writer = ExcelWriter(filename)
@@ -2620,13 +2639,3 @@ class OdkParser():
         except Exception as e:
             terminal.tprint(str(e), 'fail')
             return True, 'There was an error while saving the group settings'
-
-    def refresh_view_data(self, request):
-        try:
-            added_records = 0
-
-            return False, 'Successful refresh. %d records were added to the saved view' % added_records
-        except Exception as e:
-            terminal.tprint(str(e), 'fail')
-            sentry.captureException()
-            return True, 'Refresh Failed! %s' % str(e)
